将透视部分移到report文件夹里
