+++ nmc_verification/nmc_vf_product/__init__.py
@@ -1,4 +1,3 @@
 
 from . import yes_or_no
 from .yes_or_no import *
-from . import perspective
