+++ nmc_verification/nmc_vf_report/perspective/middle_veri_set.py
